Improve this to look more like a java project
